Fix Dockerfile and add it to CI. (#137)
We updated our Go version in 270293, but since we weren't doing Docker
builds in CI, we didn't realize it was broken. This change updates the
base image for the Dockerfile and adds a basic build to CI that just
throws away the output.

Also fixed some minor warnings that Docker was complaining about in the
Dockerfile (e.g. ENV key=value instead of ENV key value).