[remotesigning] Split up sighash and destination validation (#141)
A small refactor to split up different validation into separate objects.
Currently our `HashValidator` validates both the transaction sighash and
the output scripts of L1 wallet transactions. In an upcoming PR, I'd
like to remove the xpubs parameter from the validation interface, and
just pass the information needed to validate scripts into the specific
validator that needs it. We also redundantly validated the sighash in
the function to validate scripts and this PR removes that as well.