[remotesigning] Don't recognize L1 wallet signing job as PSBT (#144)
When we added L1 wallet signing job validation, we forgot to update how
we recognize whether a signing job is for a force close claim or not.
This would mean some L1 wallet signing jobs would fail because we'd try
to calculate the sighash as if it were a PSBT, which it's not. This is
fixed here by checking the end of the derivation *plus* that it's not an
L1 wallet signing job.