[remotesigning] Add L1 wallet flag to validate force closure claim destinations
If our node has an L1 wallet enabled, we want to validate that when we
claim an output from a force closure, that we are sending the funds to
an L1 wallet address.

This PR adds an environment variable flag to configure our destination
script validator to check the output scripts for these signing jobs.