[remotesigning] Validate force closure claims to L1 wallet (#146)
When a user has an L1 wallet, claims on a commitment tx on a force
closure should be sent to an address from this wallet. We would like to
validate that when we get a signing job for one of these claims, that
the destination is this wallet.

To do this, this PR:
- Adds an environment variable flag to enable this validation
- Refactors our `ValidateScript` function which takes a signing job,
deserializes its bitcoin tx, and validates an expected change address.
Now instead, we deserialize the tx from the signing job in a method on
the job itself, then have two functions `ValidateChangeScript` and
`ValidateOutputScript`.
- Changes the logic in our `DestinationValidator` to now validate both
in the right scenarios

Note: our normal L1 wallet validation only applies to withdrawals, and
only validates the change address.